+++ lib/ui/canvas/CanvasCourseAnnouncementsPanel.dart
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+import 'dart:collection';
+
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
@@ -36,12 +38,16 @@ class CanvasCourseAnnouncementsPanel extends StatefulWidget {
 }
 
 class _CanvasCourseAnnouncementsPanelState extends State<CanvasCourseAnnouncementsPanel> {
-  List<CanvasDiscussionTopic>? _announcements;
+  Map<int, List<CanvasDiscussionTopic>?>? _courseAnnouncementsMap;
+  List<CanvasCourse>? _courses;
+  int? _selectedCourseId;
   int _loadingProgress = 0;
 
   @override
   void initState() {
     super.initState();
+    _selectedCourseId = widget.courseId;
+    _loadCourses();
     _loadAnnouncements();
   }
 
@@ -58,18 +64,23 @@ class _CanvasCourseAnnouncementsPanelState extends State<CanvasCourseAnnouncemen
   }
 
   Widget _buildContent() {
+    late Widget contentWidget;
     if (_isLoading) {
-      return _buildLoadingContent();
-    }
-    if (_announcements != null) {
-      if (_announcements!.isNotEmpty) {
-        return _buildAnnouncementsContent();
-      } else {
-        return _buildEmptyContent();
-      }
+      contentWidget = _buildLoadingContent();
     } else {
-      return _buildErrorContent();
+      if (_courseAnnouncementsMap != null) {
+        if (_courseAnnouncementsMap!.values.isNotEmpty) {
+          contentWidget = _buildAnnouncementsContent();
+        } else {
+          contentWidget = _buildEmptyContent();
+        }
+      } else {
+        contentWidget = _buildErrorContent();
+      }
     }
+
+    return Column(
+        children: [Padding(padding: EdgeInsets.all(16), child: _buildCourseDropDownWidget()), Expanded(child: contentWidget)]);
   }
 
   Widget _buildLoadingContent() {
@@ -78,24 +89,41 @@ class _CanvasCourseAnnouncementsPanelState extends State<CanvasCourseAnnouncemen
 
   Widget _buildErrorContent() {
     return Center(
-        child: Padding(padding: EdgeInsets.symmetric(horizontal: 28), child: Text(Localization().getStringEx('panel.canvas_announcements.load.failed.error.msg', 'Failed to load announcements. Please, try again later.'),
-            textAlign: TextAlign.center, style: TextStyle(color: Styles().colors!.fillColorPrimary, fontSize: 18))));
+        child: Padding(
+            padding: EdgeInsets.symmetric(horizontal: 28),
+            child: Text(
+                Localization().getStringEx(
+                    'panel.canvas_announcements.load.failed.error.msg', 'Failed to load announcements. Please, try again later.'),
+                textAlign: TextAlign.center,
+                style: TextStyle(color: Styles().colors!.fillColorPrimary, fontSize: 18))));
   }
 
   Widget _buildEmptyContent() {
     return Center(
-        child: Padding(padding: EdgeInsets.symmetric(horizontal: 28), child: Text(Localization().getStringEx('panel.canvas_announcements.empty.msg', 'There are no announcements for this course.'),
-            textAlign: TextAlign.center, style: TextStyle(color: Styles().colors!.fillColorPrimary, fontSize: 18))));
+        child: Padding(
+            padding: EdgeInsets.symmetric(horizontal: 28),
+            child: Text(Localization().getStringEx('panel.canvas_announcements.empty.msg', 'There are no announcements.'),
+                textAlign: TextAlign.center, style: TextStyle(color: Styles().colors!.fillColorPrimary, fontSize: 18))));
   }
 
   Widget _buildAnnouncementsContent() {
-    if (CollectionUtils.isEmpty(_announcements)) {
+    if (CollectionUtils.isEmpty(_courseAnnouncementsMap?.values)) {
       return Container();
     }
 
     List<Widget> annoucementWidgetList = [];
-    for (CanvasDiscussionTopic announcement in _announcements!) {
-      annoucementWidgetList.add(_buildAnnouncementItem(announcement));
+    bool showCourseLabel = (_selectedCourseId == null);
+    for (int courseId in _courseAnnouncementsMap!.keys) {
+      CanvasCourse? course = _getCurrentCourse(courseId: courseId);
+      if (course != null) {
+        if (showCourseLabel) {
+          annoucementWidgetList.add(_buildCourseLabelWidget(course.name));
+        }
+        List<CanvasDiscussionTopic>? announcements = _courseAnnouncementsMap![course.id];
+        for (CanvasDiscussionTopic announcement in announcements!) {
+          annoucementWidgetList.add(_buildAnnouncementItem(announcement));
+        }
+      }
     }
 
     return SingleChildScrollView(
@@ -103,6 +131,22 @@ class _CanvasCourseAnnouncementsPanelState extends State<CanvasCourseAnnouncemen
         child: Padding(
             padding: EdgeInsets.only(left: 16, top: 16, right: 16),
             child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: annoucementWidgetList)));
+  }
+  
+  Widget _buildCourseLabelWidget(String? label) {
+    return Padding(
+        padding: EdgeInsets.only(bottom: 16),
+        child: Container(
+            decoration: BoxDecoration(
+                color: Styles().colors!.backgroundVariant!, border: Border.all(color: Styles().colors!.blackTransparent06!, width: 1)),
+            padding: EdgeInsets.all(10),
+            child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
+              Row(crossAxisAlignment: CrossAxisAlignment.start, children: [
+                Expanded(
+                    child: Text(StringUtils.ensureNotEmpty(label),
+                        style: TextStyle(fontSize: 18, color: Colors.black, fontFamily: Styles().fontFamilies!.bold)))
+              ])
+            ])));
   }
 
   Widget _buildAnnouncementItem(CanvasDiscussionTopic announcement) {
@@ -159,10 +203,109 @@ class _CanvasCourseAnnouncementsPanelState extends State<CanvasCourseAnnouncemen
     Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasAnnouncementDetailPanel(announcement: announcement)));
   }
 
+  Widget _buildCourseDropDownWidget() {
+    double height = MediaQuery.of(context).textScaleFactor * 62;
+    return Container(
+        height: height,
+        decoration: BoxDecoration(
+            color: Colors.white,
+            border: Border.all(color: Styles().colors!.lightGray!, width: 1),
+            borderRadius: BorderRadius.all(Radius.circular(4))),
+        child: Padding(
+            padding: EdgeInsets.only(left: 10),
+            child: DropdownButtonHideUnderline(
+                child: DropdownButton(
+                    style: TextStyle(color: Styles().colors!.textSurfaceAccent, fontSize: 16, fontFamily: Styles().fontFamilies!.bold),
+                    items: _buildCourseDropDownItems,
+                    value: _selectedCourseId,
+                    itemHeight: null,
+                    isExpanded: true,
+                    onChanged: (courseId) => _onCourseIdChanged(courseId)))));
+  }
+
+  List<DropdownMenuItem<int>>? get _buildCourseDropDownItems {
+    if(CollectionUtils.isEmpty(_courses)) {
+      return null;
+    }
+    List<DropdownMenuItem<int>> items = [];
+    CanvasCourse? currentCourse = _getCurrentCourse(courseId: widget.courseId);
+    Color textColor = Styles().colors!.textSurfaceAccent!;
+    double textFontSize = 16;
+    if (currentCourse != null) {
+      items.add(DropdownMenuItem(
+          value: currentCourse.id,
+          child: Text(StringUtils.ensureNotEmpty(currentCourse.name),
+              style: TextStyle(
+                  color: textColor,
+                  fontSize: textFontSize,
+                  fontFamily: ((_selectedCourseId == currentCourse.id) ? Styles().fontFamilies!.bold : Styles().fontFamilies!.regular)))));
+    }
+    items.add(DropdownMenuItem(
+        value: null,
+        child: Text(Localization().getStringEx('panel.canvas_announcements.all_courses.label', 'All Courses'),
+            style: TextStyle(
+                color: textColor,
+                fontSize: textFontSize,
+                fontFamily: ((_selectedCourseId == null) ? Styles().fontFamilies!.bold : Styles().fontFamilies!.regular)))));
+    return items;
+  }
+
+  CanvasCourse? _getCurrentCourse({int? courseId}) {
+    CanvasCourse? selectedCourse;
+    if (CollectionUtils.isNotEmpty(_courses)) {
+      for (CanvasCourse course in _courses!) {
+        if (course.id == courseId) {
+          selectedCourse = course;
+          break;
+        }
+      }
+    }
+    return selectedCourse;
+  }
+
+  void _onCourseIdChanged(dynamic courseId) {
+    if (_selectedCourseId != courseId) {
+      _selectedCourseId = courseId;
+      _loadAnnouncements();
+    }
+  }
+
   void _loadAnnouncements() {
+    if (_courseAnnouncementsMap != null) {
+      _courseAnnouncementsMap = null;
+    }
+    if (_selectedCourseId != null) {
+      _loadAnnouncementsForSingleCourse(_selectedCourseId!);
+    } else {
+      _loadAnnouncementsForAllCourses();
+    }
+  }
+
+  void _loadAnnouncementsForSingleCourse(int courseId) {
     _increaseProgress();
-    Canvas().loadAnnouncementsForCourse(widget.courseId).then((announcements) {
-      _announcements = announcements;
+    Canvas().loadAnnouncementsForCourse(courseId).then((announcements) {
+      if (CollectionUtils.isNotEmpty(announcements)) {
+        if (_courseAnnouncementsMap == null) {
+          _courseAnnouncementsMap = HashMap();
+        }
+        _courseAnnouncementsMap![courseId] = announcements;
+      }
+      _decreaseProgress();
+    });
+  }
+
+  void _loadAnnouncementsForAllCourses() {
+    if (CollectionUtils.isNotEmpty(_courses)) {
+      for (CanvasCourse course in _courses!) {
+        _loadAnnouncementsForSingleCourse(course.id!);
+      }
+    }
+  }
+
+  void _loadCourses() {
+    _increaseProgress();
+    Canvas().loadCourses().then((courses) {
+      _courses = courses;
       _decreaseProgress();
     });
   }